Fixed an issue in annotateCandidates causing annotations to switch organisms when a similar miRNA was found outside of the current organism of interest. This was caused by reinitiaizing the organism variable. In miRador.py, I made many changes to the configuration parsing code including the allowance of an entire folder for library tag count files to be pulled from, the user can now specify the output folder they would like to write to in order to clean up the output files created and make them easier to navigate, the code looks for the existence of all input files to prevent simple errors deep into execution, and checked for the existence of all external programs prior to reaching their actual function calls. Changes were made to the config file format, but those will be made in a separate commit

+++ miRador/annotateCandidates.py
@@ -422,10 +422,15 @@ def annotateCandidates(outputFolder, similarityDict, organism, numLibs):
 
                 # If the miRNA is known, update the image filename
                 # within the image folder to assist in quick identification
-                # with the given name in the final annotated csv file
-                os.rename("%s/images/%s_precursor.pdf" % (outputFolder, 
-                    mirName), "%s/images/%s_precursor.pdf" % (outputFolder,
-                    similarityDict[mirName]))
+                # with the given name in the final annotated csv file.
+                # But if the name with the candidate sequence does not exist,
+                # that suggests that it has already been upduated in a
+                # previous run and thus we do not need to rename the file
+                if(os.path.isfile("%s/images/%s_precursor.pdf" % \
+                        (outputFolder, mirName))):
+                    os.rename("%s/images/%s_precursor.pdf" % (outputFolder, 
+                        mirName), "%s/images/%s_precursor.pdf" % (outputFolder,
+                        similarityDict[mirName]))
 
             # If the candidate miRNA is similar to sequences in mirBase, but
             # not identical to anything in this organism, we will need to
@@ -439,6 +444,8 @@ def annotateCandidates(outputFolder, similarityDict, organism, numLibs):
                     # in miRBase, we will tag it as a member of this family
                     if(organism in organismList):
                         similarFlag = True
+                        print(mirName, similarityDict[mirName])
+                        print(organism)
                         line.append("New member of existing family")
                         line.append("%s-%s-like" % (organism, mirFamily))
 
@@ -457,11 +464,11 @@ def annotateCandidates(outputFolder, similarityDict, organism, numLibs):
                             # miRNA family that met the similarity
                             # requirement
                             toWrite = ""
-                            for organism in sorted(organismList):
-                                if(organism == organismList[-1]):
-                                    toWrite += "%s" % organism
+                            for similarOrganism in sorted(organismList):
+                                if(similarOrganism == organismList[-1]):
+                                    toWrite += "%s" % similarOrganism
                                 else:
-                                    toWrite += "%s " % organism
+                                    toWrite += "%s " % similarOrganism
 
                             line.append(toWrite)
